requirements.txt
streamlit==1.32.2
psycopg2-binary==2.9.9
fpdf==1.7.2
